+++ src/view/pages/members/meuPerfil.js
@@ -4,6 +4,7 @@ import { membersPages } from "../../../model/adminAssets.json";
 
 import MenuMinified from "../../components/MenuMinified";
 import ProfileUpdate from "../../components/ProfileUpdate";
+import Footer from "../../components/Footer";
 
 export default function MeuPerfil() {
   const userName = "userteste";
@@ -19,6 +20,7 @@ export default function MeuPerfil() {
       <main className="row no-gutters justify-content-center">
         <ProfileUpdate />
       </main>
+      <Footer/>
     </div>
   );
 }