+++ src/view/pages/members/meusFilhosDeJaco.js
@@ -6,6 +6,7 @@ import { getUserNameFromURL } from "../../../model/library/utils";
 
 import MenuMinified from "../../components/MenuMinified";
 import JacobsSuns from "../../components/JacobsSuns";
+import Footer from "../../components/Footer";
 
 export default function MeusFilhosDeJaco() {
   const [userName, setUserName] = useState("");
@@ -14,22 +15,31 @@ export default function MeusFilhosDeJaco() {
   useEffect(() => {
     setUserName(getUserNameFromURL());
     setMyJacobSuns(["jose", "asser", "issacar", "juda", "gade"]);
-  }, [])
+  }, []);
 
   return (
-    <div id="MeusFilhosDeJaco">
-      <MenuMinified
-        pages={membersPages}
-        userName={userName}
-        bgColor="dark"
-        variantColor="dark"
-      />
-      <h2 className="memberAreaPageTitle">Meus filhos de Jacó</h2>
-      <main className="row no-gutters justify-content-center">
-        <div className="col-9" id="stickersFrame">
-          <JacobsSuns myJacobSuns={myJacobSuns}/>
+    <div id="MeusFilhosDeJaco" className="container-fluid px-0">
+      <div className="row no-gutters">
+        <div className="col-12">
+          <MenuMinified
+            pages={membersPages}
+            userName={userName}
+            bgColor="dark"
+            variantColor="dark"
+          />
+        </div>
+      </div>
+      <div className="row no-gutters">
+        <h2 className="memberAreaPageTitle col-12">Meus filhos de Jacó</h2>
+      </div>
+      <main>
+        <div className="row no-gutters justify-content-center">
+          <div className="col-9" id="stickersFrame">
+            <JacobsSuns myJacobSuns={myJacobSuns} />
+          </div>
         </div>
       </main>
+      <Footer />
     </div>
   );
 }